server/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform
requests